No pre-ApiVersion handling of failed ApiVersion requests
There has been heuristics to identify a closed connection attempt
as an ApiVersionRequest not being supported by older <0.10 brokers,
where also the behaviour between 0.8 and 0.9 brokers differ.
This has been fruitful during the transition period to newer brokers
but at this point with the vast majority of users on >=0.10 brokers the
heuristics makes it harder to detect and identify configuration issues,
such as for improperly configured security.protocol, etc.

It is now time to remove the heuristics in favour of making the
experience better for modern Kafka users.

Users still on <0.10 will need to set api.version.request=false
and broker.version.fallback=.. according to existing instructions.